+++ .github/scripts/test_issue255_update_ui_write_suppression_runtime.mjs
@@ -15,8 +15,8 @@ import { instrumentSource } from "../../tools/build-render-probe-userscript.mjs"
 const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
 const SOURCE_PATH = path.join(ROOT, "src/MissionChief_Map_Command_Toolkit.user.js");
 const BASELINE_PATH = path.join(ROOT, "docs/audits/issue-255/unchanged-update-ui.json");
-const EXPECTED_VERSION = "10.3.1";
-const EXPECTED_SHA = "b318a72d2bb20c03db75406ed76ecfcf4500a739e32ce9de91490dc02bd5e829";
+const EXPECTED_VERSION = "10.3.2";
+const EXPECTED_SHA = "4b9b2f1617114a8e7e42c86c6845cf32ae3fcfc30453d6a97a9bb6bbeacc3c55";
 const REPEATS = 25;
 const HELPER_NAMES = ["normaliseDiscordReportComplexity", "discordReportComplexityAtLeast", "updateUiToggleClass", "updateUiSetStyleProperty", "updateUiSetAttribute", "updateUiSetDataset", "updateUiSetProperty", "updateUiSetText", "commandInterfaceApplySearch", "updateCommandInterfaceHeader"];
